+++ src/main/java/http/request/RequestHeaders.java
@@ -1,38 +1,38 @@
-package http;
+package http.request;
 
 import static java.util.stream.Collectors.*;
 
 import java.util.List;
 import java.util.Map;
 
-public class Headers {
+public class RequestHeaders {
     private static final String DELIMITER = ": ";
     private static final int KEY = 0;
     private static final int VALUE = 1;
 
-    private final Map<String, String> httpHeaders;
+    private final Map<String, String> requestHeaders;
 
-    private Headers(Map<String, String> httpHeaders) {
-        this.httpHeaders = httpHeaders;
+    private RequestHeaders(Map<String, String> requestHeaders) {
+        this.requestHeaders = requestHeaders;
     }
 
-    public static Headers from(List<String> httpHeaders) {
-        return httpHeaders.stream()
+    public static RequestHeaders from(List<String> requestHeaders) {
+        return requestHeaders.stream()
                 .map(line -> line.split(DELIMITER))
                 .collect(
                         collectingAndThen(
                                 toMap(
                                         token -> token[KEY],
                                         token -> token[VALUE]),
-                                Headers::new)
+                                RequestHeaders::new)
                 );
     }
 
     public boolean contains(String attribute) {
-        return httpHeaders.containsKey(attribute);
+        return requestHeaders.containsKey(attribute);
     }
 
     public String getAttribute(String attribute) {
-        return httpHeaders.get(attribute);
+        return requestHeaders.get(attribute);
     }
 }
